refactor state
