+++ lib/features/transaction/bloc/transaction_state.dart
